split codex drivers into process, rpc client, and events translator
