+++ go/harness/claude/README.md
@@ -13,9 +13,9 @@ harness.
 | Path | Look here for |
 | --- | --- |
 | [`../../core/v2/translator/claude`](../../core/v2/translator/claude) | Translating `Harness`, `AgentTemplate`, model, MCP, plugin, and Secret inputs into a runtime revision and warnings |
-| [`config`](config) | The versioned JSON contract shared by the compiler and runtime, including defaults and reserved environment variables |
+| [`config/config.go`](config/config.go) | The versioned JSON contract shared by the compiler and runtime, including defaults and reserved environment variables |
 | [`cmd/main.go`](cmd/main.go) | Actor startup, environment inputs, Claude version validation, continuation-store wiring, and private A2A startup |
-| [`internal/adapter`](internal/adapter) | Materializing Claude home, skills, MCP config, and ephemeral provider credentials |
+| [`internal/adapter/adapter.go`](internal/adapter/adapter.go) | Materializing Claude home, skills, MCP config, and ephemeral provider credentials |
 | [`internal/driver`](internal/driver) | Claude CLI arguments, stream-JSON parsing, runtime-event translation, cancellation, and process supervision |
 
 ## Working
