+++ go/harness/codex/README.md
@@ -14,7 +14,7 @@ harness.
 | Path | Look here for |
 | --- | --- |
 | [`../../core/v2/translator/codex`](../../core/v2/translator/codex) | Translating `Harness`, `AgentTemplate`, model, MCP, plugin, and Secret inputs into a runtime revision |
-| [`config`](config) | The versioned JSON contract shared by the compiler and runtime, including defaults and validation |
+| [`config/config.go`](config/config.go) | The versioned JSON contract shared by the compiler and runtime, including defaults and validation |
 | [`cmd/main.go`](cmd/main.go) | Actor startup, environment inputs, Codex version validation, continuation-store wiring, and private A2A startup |
 | [`internal/adapter`](internal/adapter) | Materializing `CODEX_HOME`, native TOML, shared agents, skills, and MCP configuration |
 | [`internal/driver`](internal/driver) | App Server lifecycle, JSON-RPC framing, event translation, thread resume, cancellation, and process supervision |
